refactor(feed): extract stories pipeline + add pagination and rate limiting
- Move all DB queries and narrative assembly logic to pipeline.ts
- Rewrite route.ts as wiring-only (~75 lines): publicRateLimit →
  getCacheOrFetch → enrichStoriesForUser → slice → headers → return
- Add offset/limit pagination; route now returns hasMore and total
  so useStoriesFeed.loadMore() can properly paginate
- Share one cached snapshot across all users; per-user like/share
  enrichment applied post-cache via enrichStoriesForUser()
- Rate limiting and X-RateLimit-* headers now match for-you pattern
