+++ apps/web/src/app/api/feed/stories/route.ts
@@ -1,659 +1,84 @@
-/**
- * Stories Feed API
- *
- * @route GET /api/feed/stories — daily-topic-focused narrative feed
- *
- * Returns narrative stories sorted by score, with stories matching today's
- * daily topic boosted to the top. Replaces the "Hot" tab with a coherent,
- * topic-driven experience per BAB-134 (one topic per day) and BAB-199
- * (story-by-story narrative sorting).
- *
- * Story Scoring:
- *   storyScore = calculateStoryScore(...) * arcMultiplier * resolutionBoost * topicBoost
- *   topicBoost = 2.0 when story.topicKey === today's topicKey, else 1.0
- *
- * @openapi
- * /api/feed/stories:
- *   get:
- *     tags:
- *       - Feed
- *     summary: Get stories feed
- *     description: Returns narrative stories for today's daily topic, sorted by narrative score.
- *     responses:
- *       200:
- *         description: Stories feed with today's topic metadata
- */
-
 import {
+  addPublicReadHeaders,
   getCacheOrFetch,
-  optionalAuth,
+  publicRateLimit,
   successResponse,
   withErrorHandling,
 } from '@babylon/api';
-import {
-  and,
-  arcStates,
-  db,
-  desc,
-  eq,
-  gte,
-  inArray,
-  isNull,
-  lte,
-  markets,
-  posts,
-  questions,
-  reactions,
-  shares,
-  sql,
-  users,
-} from '@babylon/db';
-import { dailyTopicService, StaticDataRegistry } from '@babylon/engine';
-import type {
-  ArcStateType,
-  NarrativePost,
-  NarrativeStory,
-} from '@babylon/shared';
-import { logger } from '@babylon/shared';
 import type { NextRequest } from 'next/server';
 import {
-  calculateArcStateMultiplier,
-  calculateResolutionBoost,
-  calculateStoryScore,
-} from '@/app/api/feed/narrative/scoring';
+  buildStoriesFeed,
+  enrichStoriesForUser,
+  type StoriesPipelineResult,
+} from './pipeline';
 
-const MAX_CANDIDATE_POSTS = 500;
-const MAX_STANDALONE_POSTS = 20;
-const MIN_STANDALONE_SCORE = 0.05;
-const TWELVE_HOURS_MS = 12 * 60 * 60 * 1000;
-// Stories matching today's topic are surfaced first
-const TOPIC_MATCH_MULTIPLIER = 2.0;
-const GENERAL_STORY_KEY = '__general__';
+const PAGE_SIZE = 20;
+const CACHE_TTL_S = 60;
 
-function toISOStringStrict(
-  date: Date | string | null | undefined,
-  fieldName: string,
-  postId: string
-): string {
-  if (date === null || date === undefined) {
-    logger.warn(
-      `Null/undefined ${fieldName} for post ${postId}`,
-      { postId },
-      'StoriesFeedAPI'
-    );
-    return new Date().toISOString();
-  }
-  if (date instanceof Date) {
-    if (isNaN(date.getTime())) {
-      logger.warn(
-        `Invalid Date for ${fieldName} on post ${postId}`,
-        { postId },
-        'StoriesFeedAPI'
-      );
-      return new Date().toISOString();
-    }
-    return date.toISOString();
-  }
-  const parsed = new Date(date);
-  if (!isNaN(parsed.getTime())) return parsed.toISOString();
-  logger.warn(
-    `Unparseable ${fieldName} "${date}" for post ${postId}`,
-    { postId },
-    'StoriesFeedAPI'
-  );
-  return new Date().toISOString();
-}
-
-interface QuestionMeta {
-  title: string;
-  status: string;
-  arcState: ArcStateType | null;
-  resolutionDate: Date;
-  topicKey: string | null;
-}
-
-interface CachedResult {
-  stories: NarrativeStory[];
-  postIds: string[];
-  topic: { topicKey: string; topicLabel: string; summary: string } | null;
-}
+// Stories are globally scored (not user-personalised), so all users share one
+// cached snapshot. Per-user like/share state is applied after cache lookup.
+const CACHE_KEY = 'feed:stories:v1:current';
 
 export const GET = withErrorHandling(async (request: NextRequest) => {
-  const user = await optionalAuth(request).catch((err) => {
-    logger.debug('optionalAuth failed', { error: err }, 'StoriesFeedAPI');
-    return null;
-  });
+  const {
+    error: rateLimitError,
+    user,
+    rateLimitInfo,
+  } = await publicRateLimit(request, 'read');
+  if (rateLimitError) {
+    return rateLimitError;
+  }
 
-  // Resolve today's topic before cache so the cache key is topic-aware
-  const todaysTopic = await dailyTopicService.getCurrentTopic().catch((err) => {
-    logger.warn('Failed to get daily topic', { error: err }, 'StoriesFeedAPI');
-    return null;
-  });
+  const { searchParams } = request.nextUrl;
+  const rawOffset = Number(searchParams.get('offset') ?? 0);
+  const rawLimit = Number(searchParams.get('limit') ?? PAGE_SIZE);
+  const offset = Number.isFinite(rawOffset) ? Math.max(0, rawOffset) : 0;
+  const limit = Number.isFinite(rawLimit)
+    ? Math.min(PAGE_SIZE, Math.max(1, rawLimit))
+    : PAGE_SIZE;
 
-  const cacheKey = `feed:stories:v1:${todaysTopic?.topicKey ?? 'none'}`;
+  const userId = user?.userId ?? null;
 
-  const result = await getCacheOrFetch<CachedResult>(
-    cacheKey,
-    async () => {
-      const now = new Date();
-      const cutoff = new Date(now.getTime() - TWELVE_HOURS_MS);
-
-      const recentPosts = await db
-        .select({
-          id: posts.id,
-          content: posts.content,
-          authorId: posts.authorId,
-          timestamp: posts.timestamp,
-          type: posts.type,
-          articleTitle: posts.articleTitle,
-          fullContent: posts.fullContent,
-          category: posts.category,
-          imageUrl: posts.imageUrl,
-          relatedQuestion: posts.relatedQuestion,
-          originalPostId: posts.originalPostId,
-        })
-        .from(posts)
-        .where(
-          and(
-            isNull(posts.deletedAt),
-            gte(posts.timestamp, cutoff),
-            lte(posts.timestamp, now),
-            isNull(posts.commentOnPostId),
-            isNull(posts.parentCommentId)
-          )
-        )
-        .orderBy(desc(posts.timestamp))
-        .limit(MAX_CANDIDATE_POSTS);
-
-      const topicMeta = todaysTopic
-        ? {
-            topicKey: todaysTopic.topicKey,
-            topicLabel: todaysTopic.topicLabel,
-            summary: todaysTopic.summary,
-          }
-        : null;
-
-      if (recentPosts.length === 0) {
-        return { stories: [], postIds: [], topic: topicMeta };
-      }
-
-      const postIds = recentPosts.map((p) => p.id);
-
-      // Single CTE for all engagement counts — avoids N separate round-trips
-      const postIdsArray = sql`ARRAY[${sql.join(
-        postIds.map((id) => sql`${id}`),
-        sql`, `
-      )}]::text[]`;
-
-      const engagementRows = await db.execute(sql`
-        WITH
-        target_posts AS (
-          SELECT unnest(${postIdsArray}) AS post_id
-        ),
-        reaction_counts AS (
-          SELECT r."postId" AS post_id, COUNT(*) AS count
-          FROM "Reaction" r
-          INNER JOIN target_posts tp ON r."postId" = tp.post_id
-          WHERE r.type = 'like'
-          GROUP BY r."postId"
-        ),
-        comment_counts AS (
-          SELECT c."postId" AS post_id, COUNT(*) AS count
-          FROM "Comment" c
-          INNER JOIN target_posts tp ON c."postId" = tp.post_id
-          WHERE c."deletedAt" IS NULL
-          GROUP BY c."postId"
-        ),
-        share_counts AS (
-          SELECT s."postId" AS post_id, COUNT(*) AS count
-          FROM "Share" s
-          INNER JOIN target_posts tp ON s."postId" = tp.post_id
-          GROUP BY s."postId"
-        )
-        SELECT
-          tp.post_id,
-          COALESCE(rc.count, 0) AS like_count,
-          COALESCE(cc.count, 0) AS comment_count,
-          COALESCE(sc.count, 0) AS share_count
-        FROM target_posts tp
-        LEFT JOIN reaction_counts rc ON tp.post_id = rc.post_id
-        LEFT JOIN comment_counts cc ON tp.post_id = cc.post_id
-        LEFT JOIN share_counts sc ON tp.post_id = sc.post_id
-      `);
-
-      const reactionMap = new Map<string, number>();
-      const commentMap = new Map<string, number>();
-      const shareMap = new Map<string, number>();
-
-      const engagementResultRows = Array.isArray(engagementRows)
-        ? (engagementRows as Record<string, unknown>[])
-        : [];
-      for (const row of engagementResultRows) {
-        const postId = String(row['post_id'] ?? '');
-        if (!postId) continue;
-        reactionMap.set(postId, Number(row['like_count'] ?? 0));
-        commentMap.set(postId, Number(row['comment_count'] ?? 0));
-        shareMap.set(postId, Number(row['share_count'] ?? 0));
-      }
-
-      const authorIds = [...new Set(recentPosts.map((p) => p.authorId))];
-      const authorUsers = await db
-        .select({
-          id: users.id,
-          username: users.username,
-          displayName: users.displayName,
-          profileImageUrl: users.profileImageUrl,
-        })
-        .from(users)
-        .where(inArray(users.id, authorIds));
-      const userMap = new Map(authorUsers.map((u) => [u.id, u]));
-
-      // Repost original posts for PostCard rendering
-      const repostOriginalIds = [
-        ...new Set(
-          recentPosts
-            .filter((p) => p.originalPostId)
-            .map((p) => p.originalPostId as string)
-        ),
-      ];
-      const originalPostMap = new Map<
-        string,
-        {
-          id: string;
-          content: string;
-          authorId: string;
-          timestamp: Date;
-          profileImageUrl: string | null;
-          username: string | null;
-          displayName: string | null;
-        }
-      >();
-      if (repostOriginalIds.length > 0) {
-        const originalRows = await db
-          .select({
-            id: posts.id,
-            content: posts.content,
-            authorId: posts.authorId,
-            timestamp: posts.timestamp,
-          })
-          .from(posts)
-          .where(inArray(posts.id, repostOriginalIds));
-        const originalAuthorIds = [
-          ...new Set(originalRows.map((r) => r.authorId)),
-        ];
-        const originalAuthorUsers =
-          originalAuthorIds.length > 0
-            ? await db
-                .select({
-                  id: users.id,
-                  username: users.username,
-                  displayName: users.displayName,
-                  profileImageUrl: users.profileImageUrl,
-                })
-                .from(users)
-                .where(inArray(users.id, originalAuthorIds))
-            : [];
-        const originalUserMap = new Map(
-          originalAuthorUsers.map((u) => [u.id, u])
-        );
-        for (const r of originalRows) {
-          const u = originalUserMap.get(r.authorId);
-          originalPostMap.set(r.id, {
-            id: r.id,
-            content: r.content,
-            authorId: r.authorId,
-            timestamp: r.timestamp,
-            profileImageUrl: u?.profileImageUrl ?? null,
-            username: u?.username ?? null,
-            displayName: u?.displayName ?? null,
-          });
-        }
-      }
-
-      // Question metadata — includes topicKey so we can apply topic boost
-      const questionNumbers = [
-        ...new Set(
-          recentPosts
-            .map((p) => p.relatedQuestion)
-            .filter((q): q is number => q !== null && q !== undefined)
-        ),
-      ];
-      const questionMetaMap = new Map<number, QuestionMeta>();
-      if (questionNumbers.length > 0) {
-        const rows = await db
-          .select({
-            questionNumber: questions.questionNumber,
-            text: questions.text,
-            status: questions.status,
-            arcState: arcStates.currentState,
-            resolutionDate: questions.resolutionDate,
-            topicKey: questions.topicKey,
-          })
-          .from(questions)
-          .leftJoin(arcStates, eq(arcStates.questionId, questions.id))
-          .where(inArray(questions.questionNumber, questionNumbers));
-        rows.forEach((q) =>
-          questionMetaMap.set(q.questionNumber, {
-            title: q.text,
-            status: q.status ?? 'active',
-            arcState: (q.arcState as ArcStateType | null) ?? null,
-            resolutionDate: q.resolutionDate,
-            topicKey: q.topicKey ?? null,
-          })
-        );
-      }
-
-      // Group posts into story buckets by relatedQuestion
-      const storyPostMap = new Map<string, NarrativePost[]>();
-
-      for (const post of recentPosts) {
-        const likeCount = reactionMap.get(post.id) ?? 0;
-        const commentCount = commentMap.get(post.id) ?? 0;
-        const shareCount = shareMap.get(post.id) ?? 0;
-        const timestamp = toISOStringStrict(
-          post.timestamp,
-          'timestamp',
-          post.id
-        );
-
-        const authorUser = userMap.get(post.authorId);
-        const actorRecord = StaticDataRegistry.getActor(post.authorId);
-        const orgRecord = actorRecord
-          ? null
-          : StaticDataRegistry.getOrganization(post.authorId);
-
-        let authorName = post.authorId;
-        let authorUsername: string | null = null;
-        let authorProfileImageUrl: string | null = null;
-        const authorType: 'actor' | 'news' | 'user' = actorRecord
-          ? 'actor'
-          : orgRecord
-            ? 'news'
-            : 'user';
-
-        // Filter org "NEW MARKET:" announcements — NewMarketCard is the canonical surface
-        if (
-          orgRecord &&
-          post.content.trimStart().toUpperCase().startsWith('NEW MARKET:')
-        ) {
-          continue;
-        }
-
-        if (actorRecord) {
-          authorName = actorRecord.name;
-          authorUsername = actorRecord.username ?? actorRecord.id;
-          authorProfileImageUrl = actorRecord.profileImageUrl ?? null;
-        } else if (orgRecord) {
-          authorName = orgRecord.name;
-          authorUsername = orgRecord.id;
-          authorProfileImageUrl = orgRecord.imageUrl ?? null;
-        } else if (authorUser) {
-          authorName =
-            authorUser.displayName ?? authorUser.username ?? post.authorId;
-          authorUsername = authorUser.username;
-          authorProfileImageUrl = authorUser.profileImageUrl;
-        }
-
-        const isRepost = post.type === 'repost';
-        const isQuote = isRepost && post.content !== '';
-        const originalPostData = post.originalPostId
-          ? (originalPostMap.get(post.originalPostId) ?? null)
-          : null;
-        let originalPost: NarrativePost['originalPost'] = null;
-        if (originalPostData) {
-          const origActor = StaticDataRegistry.getActor(
-            originalPostData.authorId
-          );
-          const origOrg = origActor
-            ? null
-            : StaticDataRegistry.getOrganization(originalPostData.authorId);
-          const origAuthorName =
-            origActor?.name ??
-            origOrg?.name ??
-            originalPostData.displayName ??
-            originalPostData.username ??
-            originalPostData.authorId;
-          originalPost = {
-            id: originalPostData.id,
-            content: originalPostData.content,
-            authorId: originalPostData.authorId,
-            authorName: origAuthorName,
-            authorUsername:
-              origActor?.username ??
-              origOrg?.id ??
-              originalPostData.username ??
-              null,
-            authorProfileImageUrl:
-              origActor?.profileImageUrl ??
-              origOrg?.imageUrl ??
-              originalPostData.profileImageUrl ??
-              null,
-            timestamp: toISOStringStrict(
-              originalPostData.timestamp,
-              'timestamp',
-              originalPostData.id
-            ),
-          };
-        }
-
-        const narrativePost: NarrativePost = {
-          id: post.id,
-          content: post.content,
-          fullContent: post.fullContent ?? null,
-          articleTitle: post.articleTitle ?? null,
-          category: post.category ?? null,
-          imageUrl: post.imageUrl ?? null,
-          type: post.type,
-          timestamp,
-          authorId: post.authorId,
-          authorName,
-          authorUsername,
-          authorProfileImageUrl,
-          likeCount,
-          commentCount,
-          shareCount,
-          isLiked: false,
-          isShared: false,
-          relatedQuestion: post.relatedQuestion ?? null,
-          authorType,
-          isRepost,
-          isQuote,
-          quoteComment: isQuote ? post.content : null,
-          originalPostId: post.originalPostId ?? null,
-          originalPost,
-        };
-
-        const storyKey =
-          post.relatedQuestion != null
-            ? String(post.relatedQuestion)
-            : GENERAL_STORY_KEY;
-        const bucket = storyPostMap.get(storyKey);
-        if (bucket) {
-          bucket.push(narrativePost);
-        } else {
-          storyPostMap.set(storyKey, [narrativePost]);
-        }
-      }
-
-      const todaysTopicKey = todaysTopic?.topicKey ?? null;
-      const todaysTopicLabel = todaysTopic?.topicLabel ?? null;
-      const stories: NarrativeStory[] = [];
-
-      for (const [storyKey, storyPosts] of storyPostMap) {
-        storyPosts.sort(
-          (a, b) =>
-            new Date(b.timestamp).getTime() - new Date(a.timestamp).getTime()
-        );
-
-        const newestTimestamp = new Date(storyPosts[0]!.timestamp);
-        const totalLikes = storyPosts.reduce((acc, p) => acc + p.likeCount, 0);
-        const totalComments = storyPosts.reduce(
-          (acc, p) => acc + p.commentCount,
-          0
-        );
-        const totalShares = storyPosts.reduce(
-          (acc, p) => acc + p.shareCount,
-          0
-        );
-        const questionNumber =
-          storyKey === GENERAL_STORY_KEY ? null : parseInt(storyKey, 10);
-        const meta =
-          questionNumber !== null ? questionMetaMap.get(questionNumber) : null;
-        const storyTitle =
-          meta?.title ??
-          (questionNumber !== null ? `Story #${questionNumber}` : 'General');
-        const arcState = meta?.arcState ?? null;
-        const storyTopicKey = meta?.topicKey ?? null;
-
-        // Skip resolved or expired questions
-        if (meta?.status === 'resolved') continue;
-        if (meta?.resolutionDate && meta.resolutionDate <= now) continue;
-
-        const baseScore = calculateStoryScore(
-          totalLikes,
-          totalComments,
-          totalShares,
-          storyPosts.length,
-          newestTimestamp
-        );
-        const resolutionBoost = meta?.resolutionDate
-          ? calculateResolutionBoost(meta.resolutionDate)
-          : 1.0;
-        // Stories on today's topic float to the top of the feed
-        const topicBoost =
-          todaysTopicKey && storyTopicKey === todaysTopicKey
-            ? TOPIC_MATCH_MULTIPLIER
-            : 1.0;
-        const storyScoreValue =
-          baseScore *
-          calculateArcStateMultiplier(arcState) *
-          resolutionBoost *
-          topicBoost;
-
-        stories.push({
-          storyKey,
-          storyTitle,
-          questionNumber,
-          arcState,
-          storyScore: Math.round(storyScoreValue * 10000) / 10000,
-          postCount: storyPosts.length,
-          posts: storyPosts,
-          hasUserPosition: false,
-          topicKey: storyTopicKey,
-          topicLabel:
-            storyTopicKey === todaysTopicKey ? todaysTopicLabel : null,
-        });
-      }
-
-      // Standalone general posts compete on score alongside question stories
-      const generalPosts = storyPostMap.get(GENERAL_STORY_KEY) ?? [];
-      const standalonePostCards = generalPosts
-        .map((post) => ({
-          post,
-          score: calculateStoryScore(
-            post.likeCount,
-            post.commentCount,
-            post.shareCount,
-            1,
-            new Date(post.timestamp)
-          ),
-        }))
-        .sort((a, b) => b.score - a.score)
-        .slice(0, MAX_STANDALONE_POSTS)
-        .filter(({ score }) => score >= MIN_STANDALONE_SCORE);
-
-      for (const { post, score } of standalonePostCards) {
-        const rawTitle =
-          post.articleTitle ??
-          (post.content.length > 80
-            ? post.content.slice(0, 80).replace(/\s+\S*$/, '') + '…'
-            : post.content);
-        stories.push({
-          storyKey: `post:${post.id}`,
-          storyTitle: rawTitle,
-          questionNumber: null,
-          arcState: null,
-          storyScore: Math.round(score * 10000) / 10000,
-          postCount: 1,
-          posts: [post],
-          hasUserPosition: false,
-        });
-      }
-
-      // Resolve marketId for question-backed stories (enables PredictionSparkline)
-      const storyQuestionNumbers = stories
-        .filter((s) => s.questionNumber !== null)
-        .map((s) => s.questionNumber as number);
-      if (storyQuestionNumbers.length > 0) {
-        const marketRows = await db
-          .select({
-            questionNumber: questions.questionNumber,
-            marketId: markets.id,
-          })
-          .from(questions)
-          .innerJoin(
-            markets,
-            sql`lower(trim(${markets.question})) = lower(trim(${questions.text}))`
-          )
-          .where(inArray(questions.questionNumber, storyQuestionNumbers));
-        const questionToMarket = new Map(
-          marketRows.map((r) => [r.questionNumber, r.marketId])
-        );
-        for (const story of stories) {
-          if (story.questionNumber !== null) {
-            story.marketId = questionToMarket.get(story.questionNumber) ?? null;
-          }
-        }
-      }
-
-      // Final sort: topic-boosted scores already float topic stories to top
-      stories.sort((a, b) => b.storyScore - a.storyScore);
-
-      const allPostIds = [
-        ...new Set(stories.flatMap((s) => s.posts.map((p) => p.id))),
-      ];
-
-      return { stories, postIds: allPostIds, topic: topicMeta };
-    },
-    { namespace: 'feed', ttl: 60 }
+  const fullResult = await getCacheOrFetch<StoriesPipelineResult>(
+    CACHE_KEY,
+    buildStoriesFeed,
+    { ttl: CACHE_TTL_S }
   );
 
-  // Per-user like/share enrichment (bypasses cache, same pattern as hot route)
-  if (user?.userId && result.postIds.length > 0) {
-    const [userLikes, userShares] = await Promise.all([
-      db
-        .select({ postId: reactions.postId })
-        .from(reactions)
-        .where(
-          and(
-            inArray(reactions.postId, result.postIds),
-            eq(reactions.userId, user.userId),
-            eq(reactions.type, 'like')
-          )
-        ),
-      db
-        .select({ postId: shares.postId })
-        .from(shares)
-        .where(
-          and(
-            inArray(shares.postId, result.postIds),
-            eq(shares.userId, user.userId)
-          )
-        ),
-    ]);
-    const likedPostIds = new Set(userLikes.map((l) => l.postId));
-    const sharedPostIds = new Set(userShares.map((s) => s.postId));
-    for (const story of result.stories) {
-      for (const post of story.posts) {
-        post.isLiked = likedPostIds.has(post.id);
-        post.isShared = sharedPostIds.has(post.id);
-      }
+  // Per-user like/share enrichment — applied after cache, never cached.
+  if (userId && fullResult.postIds.length > 0) {
+    await enrichStoriesForUser(fullResult.stories, fullResult.postIds, userId);
+  }
+
+  const total = fullResult.stories.length;
+  const page = fullResult.stories.slice(offset, offset + limit);
+  const hasMore = offset + limit < total;
+
+  const response = successResponse({
+    success: true,
+    topic: fullResult.topic,
+    stories: page,
+    total,
+    hasMore,
+    generatedAt: fullResult.generatedAt,
+  });
+
+  if (rateLimitInfo) {
+    if (userId) {
+      response.headers.set('Cache-Control', 'private, no-store');
+      response.headers.set('X-RateLimit-Limit', rateLimitInfo.limit.toString());
+      response.headers.set(
+        'X-RateLimit-Remaining',
+        rateLimitInfo.remaining.toString()
+      );
+      response.headers.set(
+        'X-RateLimit-Reset',
+        rateLimitInfo.resetAt.toISOString()
+      );
+    } else {
+      addPublicReadHeaders(response, rateLimitInfo);
     }
   }
 
-  return successResponse({
-    success: true,
-    topic: result.topic,
-    stories: result.stories,
-  });
+  return response;
 });
